feat(rag):  Add Milvus vector database support
-Add document partitioning and vectorization mechanism module
-Add Milvus class to implement VectorStore interface
-Add the MilvueContainer class to manage Milvus client connections
-Implemented functions such as vector embedding, document addition, deletion, and querying
-Added Milvus related test cases
-Optimized some code structures, improved scalability and maintainability

+++ src/main/java/org/salt/jlangchain/ai/chat/openai/param/OpenAIResponse.java
@@ -51,7 +51,7 @@ public class OpenAIResponse {
     public static class DataObject {
         private String object;
         private int index;
-        private List<Double> embedding;
+        private List<Float> embedding;
     }
 }
